refactor: Update Result component to display user ID
The Result component has been updated to display the user ID in the user information section. The code now includes a new div element that shows the user ID retrieved from the item object. This change enhances the information displayed to the user and improves the overall functionality of the component.