refactor: Add link to open resume in Result component
The Result component has been updated to include a link that allows users to open the resume of the user. The link is displayed in the user information section and opens the resume in a new tab when clicked. This enhancement improves the functionality and user experience of the component.